implemented users module

+++ src/users/users.controller.ts
@@ -1,34 +1,67 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, NotFoundException, InternalServerErrorException, Query, ParseIntPipe } from '@nestjs/common';
+import { ApiTags, ApiOperation, ApiBearerAuth, ApiBody } from '@nestjs/swagger';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+import { UserResponseDto } from './dto/response/user.response.dto';
+import { CreateUserDto } from './dto/request/create-user.request.dto';
+import { UpdateUserDto } from './dto/request/update-user.request.dto';
+//import { AuthGuard } from '../auth/guards/jwt-auth.guard';
 
+@ApiTags('Users')
+@ApiBearerAuth()
 @Controller('users')
+//@UseGuards(AuthGuard)
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
+  @ApiOperation({ summary: 'Crear usuario' })
+  @ApiBody({ type: CreateUserDto })
+  create(@Body() createUserDto: CreateUserDto): Promise<UserResponseDto> {
+    try{
+      return this.usersService.create(createUserDto);
+    } catch (error) {
+      throw new InternalServerErrorException(error);
+    }
   }
 
   @Get()
-  findAll() {
-    return this.usersService.findAll();
+  @ApiOperation({ summary: 'Obtener todos los usuarios' })
+  findAll(@Query('clubId', ParseIntPipe) clubId: number): Promise<UserResponseDto[]> {
+    try{
+      return this.usersService.findAll(clubId);
+    } catch (error) {
+      throw new InternalServerErrorException(error);
+    }
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(+id);
+  @ApiOperation({ summary: 'Obtener usuario por ID' })
+  findOne(@Param('id', ParseIntPipe) id: number, @Query('clubId', ParseIntPipe) clubId: number, @Query('typeId', ParseIntPipe) typeId: number): Promise<UserResponseDto> {
+    try{
+      return this.usersService.findOne({ clubId, userId: id, typeId });
+    } catch (error) {
+      throw new NotFoundException(error);
+    }
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
+  @ApiOperation({ summary: 'Actualizar usuario' })
+  @ApiBody({ type: UpdateUserDto })
+  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto): Promise<UserResponseDto> {
+    try{
+      return this.usersService.update({ clubId: updateUserDto.clubId, userId: +id, typeId: updateUserDto.typeId }, updateUserDto);
+    } catch (error) {
+      throw new InternalServerErrorException(error);
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(+id);
+  @ApiOperation({ summary: 'Eliminar usuario' })
+  remove(@Param('id', ParseIntPipe) id: number, @Query('clubId', ParseIntPipe) clubId: number, @Query('typeId', ParseIntPipe) typeId: number): Promise<UserResponseDto> {
+    try{
+      return this.usersService.remove({ clubId, userId: id, typeId });
+    } catch (error) {
+      throw new InternalServerErrorException(error);
+    }
   }
 }
